requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
typer>=0.9.0
rich>=13.4.2
